server/requirements.txt
Flask
python-dotenv
Flask-RESTful
Flask-SQLAlchemy
Flask-JWT-Extended
Marshmallow
Marshmallow-SQLAlchemy
Flask-Uploads
Authlib
Alembic